+++ src/DependencyInjection/GosWebSocketExtension.php
@@ -253,6 +253,7 @@ final class GosWebSocketExtension extends Extension implements PrependExtensionI
 
                     $definition = new ChildDefinition('gos_web_socket.periodic_ping.doctrine');
                     $definition->addArgument(new Reference($serviceRef));
+                    $definition->addArgument($pingService['interval']);
                     $definition->addTag('gos_web_socket.periodic');
 
                     $container->setDefinition('gos_web_socket.periodic_ping.doctrine.'.$serviceRef, $definition);
@@ -264,6 +265,7 @@ final class GosWebSocketExtension extends Extension implements PrependExtensionI
 
                     $definition = new ChildDefinition('gos_web_socket.periodic_ping.pdo');
                     $definition->addArgument(new Reference($serviceRef));
+                    $definition->addArgument($pingService['interval']);
                     $definition->addTag('gos_web_socket.periodic');
 
                     $container->setDefinition('gos_web_socket.periodic_ping.pdo.'.$serviceRef, $definition);