+++ src/Periodic/DoctrinePeriodicPing.php
@@ -18,14 +18,14 @@ final class DoctrinePeriodicPing implements PeriodicInterface, LoggerAwareInterf
      */
     private object $connection;
 
-    private int $timeout;
+    private int $interval;
 
     /**
      * @param Connection|PingableConnection $connection
      *
      * @throws \InvalidArgumentException if the connection is not an appropriate type
      */
-    public function __construct(object $connection, int $timeout = 20)
+    public function __construct(object $connection, int $interval = 20)
     {
         if (!($connection instanceof Connection) && !($connection instanceof PingableConnection)) {
             throw new \InvalidArgumentException(sprintf('The connection must be a subclass of %s or implement %s, %s does not fulfill these requirements.', Connection::class, PingableConnection::class, \get_class($connection)));
@@ -42,7 +42,7 @@ final class DoctrinePeriodicPing implements PeriodicInterface, LoggerAwareInterf
         }
 
         $this->connection = $connection;
-        $this->timeout = $timeout;
+        $this->interval = $interval;
     }
 
     public function tick(): void
@@ -79,13 +79,13 @@ final class DoctrinePeriodicPing implements PeriodicInterface, LoggerAwareInterf
 
     public function getTimeout(): int
     {
-        return $this->timeout;
+        return $this->interval;
     }
 
     public function setTimeout(int $timeout): void
     {
         trigger_deprecation('gos/web-socket-bundle', '3.9', '%s() is deprecated and will be removed in 4.0, set the timeout through the constructor instead.', __METHOD__);
 
-        $this->timeout = $timeout;
+        $this->interval = $timeout;
     }
 }